Use separate flag to suppress resizes
When moving a task between stacks, we temporarily suppress resize
changes because while moving, the task is still in another stack
which might get resized by the docked stack because when fetching
the stack, we might create it and thus resize the other stack.

Introduce mTemporarilyUnresizable which makes it really not resizable,
regardless of whether we are force resizing our activities.

Change-Id: Ib51163a0606106fd55f5bdeecf8e53f08add4b4b

Former-commit-id: 8202b2a5db2219c72c4895b497d256ddca216821