More multi-window fixes
- When exiting multi-window (taskOverride gets EMPTY), emulate
the same task configuration so we don't end up with unnecessary
relaunches.
- Fix flicker by not calling WM.setAppVisibility if the app is
already visible. setAppVisibility makes a dummy transition,
which sets the transformation's alpha to zero if we think that
the app is not visible, which is the case because the app is in
DRAW_PENDING state because we are waiting for it to exit the
dragResize mode, but it's really visible.

Change-Id: Ieb4f586ae86e1185b21a901c57883a1f19d58fee

Former-commit-id: 1b025a6026fc988ad849303f0df7821354b84c86