Add some View methods overrides and WebViewDelegate interfaces.
There are IME use cases for returning a different value from
getHandler() and findFocus(), in order to make InputMethodManager
communicate with a separate IME thread.  Currently, it's cumbersome to
do so in WebView because we cannot alter View-inherited methods without
a framework change.

This patch pair adds the framework-side override and adds no-op stub
methods on the Chromium side, to enable us to take this approach in the
future.

https://codereview.chromium.org/1669873003 implementing the stubs will
land first, and after it rolls in we can land this to add the interface.

Change-Id: I53ffeb8540b0cb288103c3973274795efc5c2b8e

Former-commit-id: 73a7494231c95b53f65612e9fdd9b5aef506244e