Multi packages per APK - broadcasts
This change introduces the ability to have multiple packages per
APK. The feature is currently restricted to privileged apps and
updates to such apps.

In essence the manifest can have multiple child package declarations.
A child package can declare everything an Android package can except
some tags or attributes that are not applicable and instead inherited
from the parent when needed. For example, the target SDK of the parent
applies to all children.

A child package can be updated only through the parent package.
A package with multiple child packages is installed, uninstalled
atomically - no partial installs where some child packages are not
installed.

This change ensures that we send package broadcasts for child packages
when they are updated, removed, disabled, replaced, etc.

Sample app:ag/848432

Change-Id: I25b29c98152dcad9ede4d4eb040cc897b7f93426

Former-commit-id: f4ff39c92be840d5f53c42cb02cef6b03a1ca70f