Make InputMethodInfo#isEncryptionAware() strict.
This is a follow up CL to the previous CL [1], which unintentionally
relied on whether the entire application rather than the
InputMethodService is marked to be encryption-aware.  This wasn't
obvious because right now LatinIME is entirely marked to be
encryption-aware, which probably needs to be reconsidered in subsequent
CLs.

This CL drops "encryptionAware" line from InputMethodInfo#dump() as it
is redundant.  The same information is already dumped in the
'ServiceInfo' section.

 [1] Icf921fe3661eccf4a589b08b616d05decc561356
     320810f97b9e8c23018fb43987c64d1fcb66d7d9 [formerly 69811a98f161a04af8e8ec9978c3a5efe1ea0f29]

Bug: 26279466
Change-Id: I6cea5fe5731f4a6a64ea722b854b99c5154a15ae

Former-commit-id: 37db07216e7aa7baf18b7996b79ff9028aca6fb1