Support keyboard entry of non-ASCII characters
Add characters to the keyboard mapping to allow them
to be entered into the AVD from the developer's PC
keyboard.

This mapping supports the non-ASCII keys that appear
on the "US International" keyboard. Most of the Latin-1
characters (but not all) appear on this keyboard.

This change does not remove support for any keys.

b.android.com/200042

Change-Id: I65d54eb36183382b171c388b26848941eb0ff8d9

Former-commit-id: dac57966a9ddc2e3dbd233c8b1e7b83f60e11373