WifiManager: Return empty scan list on remote exception
In the (hopefully rare) situation where there is a remote exception
from a call to getScanResults(), do not pass null back to callers.
Nobody expects this, and failures that occur as a result can distract
from the real problem (e.g., the system server crashing).  This CL
now returns an empty list.

BUG=27139097
TEST=None

Change-Id: I637eae1fa8ac1a1bc2b76dea8628359b5dec06e1

Former-commit-id: ed9297991a922b88a00c04d034c172c18d550ac1