BootReceiver: Stop using SharedPreferences.
BootReceiver was using SharedPreferences to record the timestamp that a
log file was lastly added to the dropbox. It no longer works after the
cleanup of the ContextImpl, because Context storage APIs are designed
for app data storage (as opposed to the system server). This CL switches
to writing its own xml file instead (/data/system/log-files.xml).

Bug: 26966507
Change-Id: Ife6b7544a7636b37dd239f059d1ca8c9c28f81a3

Former-commit-id: 8b231a96a147d0b897a68cc9b1d0feaca515b000