Have unified setter/getter for Secure Settings.
In order to make InputMethodManagerService encryption-aware, we are
going to introduce a new state where any read/write access to Secure
Settings from IMMS is virtualized so that we can temporarily enable
only encryption-aware IMEs until the user unlocks the device then revert
any changes made before the device enters into an unlocked state.

To do that, it would be convenient if InputMethodUtils has unified
getter/setter methods to access (Secure) settings.  In subsequent CLs we
will rely on those getter/setter methods to switch between the on-memory
data store and the actual Secure Settings.  Note that because of
multi-user support such a switch can occur multiple times.

This is still a preparation code.  Behavior change is not intended yet
in this CL.

Bug: 26279466
Change-Id: I0f79243e5cc1556764da37fa38078e075a27d42b

Former-commit-id: 8752367042cb690f78953557433c16ac77eeea45