TIF: Rename setTvInputInfo to updateTvInputInfo
As TvInputInfo includes many complex objects as fields, it is not
practical to check whether an existing TvInputInfo has actually
"changed" by using simple comparisons when a new TvInputInfo object is
set by a TV input service.
This code change renames
  - setTvInputInfo to updateTvInputInfo
  - onTvInputInfoChanged to onTvInputInfoUpdated
to indicate that the callback is mechanically invoked whenever
updateTvInputInfo method is called.

Change-Id: Ia351b3591d0ed9da962311546f0ba51158e6478e

Former-commit-id: aa5605ffee270ef8802c5d9dc8df8ce71e377f55