Fixed issue with losing window name when copying WindowManager.LayoutParams
It is possible for the WindowManager.LayoutParams set from a view/window on
the client side not to have the LayoutParams.mTitle set. This is normally
not an issue as we later set it to the package name before passing the
LayoutParams to window manager. However, we rely on the full window name
in order to identify unique windows during CTS hostside tests. We now
retain the current LayoutParams.mTitle if the LayoutParams we are copying
form didn't set one.

Bug: 26982752
Change-Id: Icc3c5937391ed1646c3777921c6f61c27ebec05d

Former-commit-id: b6e2eaddfcf78f265061fbcc3598b8ca865d6bdb