Filter components, prevent app start in safe mode.
When device is in safe mode, we're trying to keep third-party code
from running to give us a stable platform.  The ideal approach would
be to treat these apps as temporarily "uninstalled" when in safe mode,
but not all system internals are ready for this.

Instead, go back to previous behavior where we simply filtered
non-system components.  This isn't perfect, since there are still
cracks through which components can leak out (GET_ACTIVITIES, etc).

So as a last-ditch sanity effort, refuse to fork any third-party apps
while running in safe mode.

Bug: 27165374
Change-Id: I044ede02e923c499159faf59b12e79b97fe77fba

Former-commit-id: c6ca265902d41a0553d15f660647410db6a3e889