Implement lenient background check option.
Which really means, make background check much more
strict, with an option to revert to the more lenient
behavior.

In this strict version, an app can't have services
started or receive broadcasts at any point when it is
not foreground.  Also, it doesn't matter the importance
of a caller trying to start a service, it only depends
on the state of the app whose service is being started.

A new activity shell command allows you to control
whether to use the strict or lenient behavior.

Change-Id: I660df480dd55575a38c9542210210ec53e34eca0

Former-commit-id: 72bea4a1f0400c21a68974176301e034c9085f21