Fix leaking animation
Clear animation when destroying window surface.

Because we are not stepping in animations for windows if they don't
have a surface, we can't use it as a detached wallpaper target
animation.

When the surface was destroyed while animating,
winAnimator.mAnimating was still true, and never got false, because
we didn't call stepAnimation on the WindowAnimator anymore.

Thus, we were constantly animating, using a lot of battery (surface
flinger was constantly pushing frames) and impacting overall system
performance.

Bug: 27276043
Change-Id: I8b0943da0908c85740dc1113aedb573fb440f162

Former-commit-id: 02b74b9420f0b4adbec6997b950e244ad0dff56c