Call webview updating logic with the system UID when setup done.
After performing the webview setup we check whether we changed webview
implementation during the setup and, if so, we rerun the updating-logic.
This logic needs special permissions such as changing settings
(Settings.Global) which the shared relro process doesn't have,
therefore, when the shared relro process notifies us that the setup is
done we need to switch to using the system uid when running the
update-logic.

Bug: 27262336
Change-Id: I68646884e31ba55c860ff98c3f30a520beb0ef95

Former-commit-id: 275d13c2c6c2f055b41b14b1c7ba78b838f03b4d