Adding missing break statements causing crash on MSG_SET_EXTRAS.
The case block for MSG_SET_EXTRAS was missing its break statement, all
the way back to summer 2015 when that code was originally added.

When the new MSG_ON_CONNECTION_EVENT handler was added, it introduced a
new potential fall-through from MSG_SET_EXTRAS (in addition to the fact it
too was missing a break statement).

Bug: 27338441
Change-Id: I2df2bc78c4ceb32e6245888eed18c2d8834e629b

Former-commit-id: 86c9fb41736f5848acf7baf71ebfe39df204ad4b