Force cancel pending RemotePrintDocument commands when the PrintActivity exits
Otherwise the print-client app can prevent the print activity from exiting as
we wait for the command to finish. As this is a violation of the
remote-print-document contract we deal with this similar to a command failure.
As the PrintActivity is exiting we don't further send any commands to the
print-client.

Bug: 27207751
Change-Id: Ieacb16786112f217ca5b8031bfb597598e28dc3d

Former-commit-id: 645a3e1ca6ce26da52fa158ca159a31bd38de77d