SensorEventListener: Clarify documentation.
"onSensorChanged" is really misnamed (perhaps we should deprecate
for a better name, but that's beyond the scope of this CL).  We
will still call this method with identical sensor values if it
has a newer timestamp.  That is different than the
"onAccuracyChanged" behavior.

We expand some of this documentation to clarify this point to users.

Change-Id: I95fa52574ebac3d20a2354b7b6dcf1bdf513448d

Former-commit-id: cc0a4820ebaaf07a4643d43d7ce25907ca78d13b