Fixed a bug where notifications could stay dark
The check whether dark changed was always done
on the notification itself, ignoring that the
contentviews actually also have a check whether it
changed, leading to potential problems that this state
became out of date. In addition did we assume that whenever
an update comes in the text becomes normal, which is
not true for an inline update. This hopefully fixes things once
and for all! I said once and for all!

Bug: 27147160
Change-Id: Idd68be8da0f648cf79102544b038065618c1d0ff

Former-commit-id: b88b98477ae725092c27dfe52f7fca3d4f29913e