SettingsLib: Always update preference state.
In case a RestrictedPreference can also be disabled
for other reasons than device admin, the state
of RestrictedPreferenceHelper may not be up-to-date
with the actual preference state. For example, the
"Android Beam" checkbox can be disabled by device
policy, but it can also be disabled by Settings itself
because NFC was turned off by the user.

To fix that, always update the Preference state.

Bug: 26907006
Change-Id: I27cde70beb82721dd4d423943a9898e022df8862

Former-commit-id: 6f02f56f52248a3129da6072ed692591cc1074f6