DO NOT MERGE ANYWHERE Add way to query for supported Bluetooth profiles.
Currently there is no way to get the profiles supported by the Bluetooth
adapter. Asking for a profile proxy of an unsupported profile does not
fail and can lead to code indefinitely waiting for the proxy response. This
new code will allow for checking the supported profiles before asking for
the proxies.

Bug: 26451648
Change-Id: Ie209f08058205eea39513c03b626788f5fed3293
Signed-off-by: Bryce Lee <brycelee@google.com>

Former-commit-id: 7c2539e2a50ec3dcdb1c9420f9215bba438b21b6