Proper grant default permissions to default SMS and Phone
We grant default permissions to the default SMS and
Phone apps when they are selected as such to ensure
they can do their job. We don't grant default permissions
to apps if the version on the system image does not
declare them. This is correct for default grants on
first boot or a new user creation. This is a problem
for default Phone and SMS as we want to grant them
the permission as a result of a deliberate user action
and therefore should not care if the version on the
system image declares the permission.

As a result if an SMS app that ships as a stub to reduce
image size is later updated to the full version and
made the default SMS it would not get the default grants
while an installed third-party app that did not ship
on the system image would get the grants.

This change ensures the default SMS and Phone get
their default grants regardless whether the verion on
the system image declares the permissions.

bug:25288760

Change-Id: I2fecf5cbce4ebabb145a3f29abb555bcb65d54d6

Former-commit-id: 80d78dc92bed0873d03e357e025f994af1217577