Ignore the checksum when comparing RAs for equality.
When comparing two RAs for equality, Ra#matches intentionally
ignores lifetimes. It should ignore the checksum instead, because
otherwise-identical packets with different lifetimes will almost
certainly have different checksums.

In order to avoid the generated APF program also matching the
checksum, we treat the checksum as if it were a lifetime, but
take care not to use its value when calculating mMinLifetime or
when checking the lifetimes in the APF program.

Bug: 27595799
Bug: 26238573
Change-Id: I526fdc17e99803a1ddec6275a3c542014434c429

Former-commit-id: e6dfc1736b3e37ad2ea0e9005cab36432de196b2