docs: Removing the permissions update page
Per bug b/27487208 , we're removing this page because there isn't
quite enough content. The material that was here will be added to the
N "Behavior Changes" page *before* Preview 1 ships. (That page is
currently in flux so I can't add it just now.)

If any of you have comments on either of the topics that were covered
here, please make comments on this CL—we'll make sure the comments
are addressed when the material is moved into behavior-changes.jd.

[Work on permissions-change doc continued here from the previous CL
(http://ag/873827/) which we submitted in the run-up to Preview 1.]

bug: 27192015
bug: 27487208
Change-Id: If637b62203192ad0d6a50b99a207ef075a453c6f

Former-commit-id: 5728bb55fc9563ebb66f16453255f042d6f66872