Redraw the UI when a print page preview is updated
When switching between two printers with different margins but with the
same paper size we were missing updates of the print preview. This
change triggers these updates.

The issue was the we reuse the a bitmap when a page is rewritten. This
causes View.setBackground(drawable) to do nothing as the old
background is the same as the new one. This function only checks the
background drawable object reference, not the content. Hence we need to
add a call that notifies the listeners to the changes on the content of
the drawable.

Bug: 26961677
Change-Id: Id8b9be59c9ca7220f40ac59168f78bfdcccb2761

Former-commit-id: 8eb451597d5f795c11e96663232ec58a2b1fc60f