[Sensor] Add a constant to match native header
This CL add private sensor type base constant to java file to match
C header file.

It is defined that all vendor sensor types have to be >= 0x10000.
There is a constant in sensor.h for reference. However, there is
no such constant defined in Sensor.java before. This CL adds it.

Change-Id: Ia8f8c06583c62879e51779815051c346f1a9cd14

Former-commit-id: 1795e14dab0367fdd77f2c1eee6b4fd39d1c5902