Switch to SystemService lifecycle in TSMS.
Like we did this for InputMethodManagerService [1],
TextServicesManagerService (TSMS) needs to be recognized by
SystemServiceManager with SystemService lifecycle mechanism so that we
can receive SystemService#onUnlockUser() event, which is necessary to
make TSMS encryption-aware.

As a preparation, with this CL we only does mechcanical migration to
SystemService lifecycle mechanism in TSMS.  Hence no user-visible
behavior change should occur.

  [1]: Ic17667df60b30e5355b61a3601ad27a000cab3a3
       b3a24739b0f634b9a679c3439a2e6af7f1e8b5b9 [formerly 1e33dc8fdf3f722ecd32cc586b2a9515de24a242]

Bug: 27456430
Change-Id: Ib3cc799d384f259b4fa3f5295f2da198df015eb8

Former-commit-id: a6a152e7de2b6db73474620c1eccda1ebe2eee9b