Add copy-on-write mode to TextServicesSettings.
This ports my previous CL for InputMethodManagerService [1] to
TextServicesSettings (TSMS).

This is a preparation for File-Based Encryption (FBE) support in TSMS.

With this CL, we start switching TextServicesSettings from copy-on-write
mode to real mode and start rebuilding the list of available spell
checkers when the system becomes ready.  Anything else should be the
same.  See the previous commit message [1] for details.

  [1]: I9c6f9bb3d51174198e5f73588637f87ea0d90e11
       8cfc235891b88543d8f7dfcae953f9c03a11c8c1 [formerly 68645a638ad1bfb734b2b0f56b17fe206bb891c5]

Bug: 27456430
Change-Id: Ie3d61458648df469abe149b7aaad8087c531a675

Former-commit-id: 49ed14017697f8f19b8d49d55462713396a0588e