Fixed visibility of home stack behind translucent app activity.
A stack is not considered translucent if their top activity is over the
home stack (i.e the home stack should be visible behind the activity).
However, the home stack will not be made visible behind this stack since
the stack is not translucent.
We now make the stack transulcent if the top activity is over the home
stack and the home stack is directly behind this stack.

Also, cleaned up the isStackTranslucent code some to be more readable.

Bug: 27566650
Change-Id: I4d2fe842d26547ba743ca82a456499038a1eb389

Former-commit-id: 56d75cf04df160438dc61fb3e1ca916ea9afc61a