Add a "Do not ask again" checkbox.
When an app request access to a scoped directory which the user already
denied access, it will display a "Do not ask again" checkbox; if the
user checks that option, further requests will be automatically
rejected.

The history of denials is stored in the shared property file.

The UI is not polished yet, the style will be fixed in a future change.

BUG: 26750152

Change-Id: I181923adfb6a1c7c1c17e305d6838314280417fc

Former-commit-id: adccb9950ee3357a7c35283235c2b01500c43aec