docs: Removing ACTION_OPEN_EXTERNAL_DIRECTORY from preview docs
ACTION_OPEN_EXTERNAL_DIRECTORY was formerly a new Intent in N
that was used to make a Scoped Directory Access request, however
it was removed and encapsulated into a new StorageVolume class.
Since the Intent no longer exists, I'm removing it completely
from the Permissions Changes section of the Behavior Changes page.

I've also grep'd through the rest of the preview docs for
ACTION_OPEN_EXTERNAL_DIRECTORY and did not find any other mention
of it that needed to be removed.

Preview stage specified in CL comment.

Bug: 27581845
Change-Id: I2ab1a38e64829d899d14ec1fc4e8eca5a5ac033d

Former-commit-id: 8f6e53f88aedc44610d27945f3b2fa2f9c4ffb65