Use LocaleUtils#filterByLanguage for non-keyboard subtypes.
With this CL, we expand the target of Bug 27129703 and Bug 27348943 to
non-keyboard subtypes.

Suppose there is a handwriting IME (mode == "handwriting") that supports
the following 5 subtypes.
  - en-US
  - en-GB
  - fr
  - sr-Cyrl
  - sr-Latn

Also suppose the system languages are configured as follows.
  1. sr-Latn-RS
  2. ja-JP
  3. fr-FR
  4. en-GB
  5. en-US

In this case we want to enable [sr-Latn, fr, en-GB] by default when
"use system language" is checked in the subtype enabler.

See previous commits [1][2] about how we addressed those issues for
keyboard subtypes.

  [1]: Iaf179d60c12b9a98b4f097e2449471c4184e049b
       895b22114d0b75dda7dec24aaa64158a3bbb0dad [formerly e985c240e3feb62ea38d5b4e386be083ca0f215b]
  [2]: I8fc774154f5175abff2f16e8f12a4847bf5f5b7c
       c57ba4c0211e60b998321502d699865a621fdd45 [formerly 072a95a3094af2ced4f009ad62c4553c28e3f830]

Bug: 27560993
Change-Id: I416b5671602c550805ed6267dd210968aa1de83c

Former-commit-id: 238faad984786f0227fe6c0cf6ea2a5ecff3a4bc