ExifInterface: ignore exceptions when getting wrong images
In order to maintain compatibility with the old versions of
ExifInterface, non-JPEG image files, for instance, PNG, GIF and so on,
need to be handled by generating an empty result rather than throwing
an exception in the constructor. And corrupted JPG files also should be
handled in the same way.

Bug: 27587980
Change-Id: If33b63c683bfb672999d1abd370a8edf29932ac1

Former-commit-id: 1043f56185ee4abb0246ebe44eea7e391802d4fa