Fix for JobScheduler jobs not running
If the device is powered off when a job is
supposed to run, its delay and deadline both
are set to 0 on boot. If the deadlineExpired alarm
runs first, the job is discarded from
TimeController's tracked jobs list. The delay
constraint is never set to true and the job
would never be executed. This behavior is fixed
here.

Dump modified to print jobId and uId instead of
hash to make it more meaningful.

Bug: 27619643
Change-Id: If03e3d6f194f15fe43b697dcb4198137fa7ed646

Former-commit-id: 0529dac587b1361eac359737a5f10972bce5fd0d