Apply stack crop to docked stack during animation
If device goes to sleep with docked stack visible, the docked stack
will be minimized after unlock. Due to "force hide enter" animation
and the entering animation of the app itself, the app will be animating
for the initial few frames. Not applying the crop causes the app to
be seen in full briefly, then minimized.

Docked divider doesn't use a final stack bounds but changes stack
bounds continously during docked stack resize, so we can apply the
bounds even during animation.

bug: 27592767
Change-Id: I12f4eb46a95c9a7dd2147b68c362e17609a8b410

Former-commit-id: 477e720696970d8ab3cc4e6569022f1db0272d43