FloatingToolbar: Hide panels when they're invisible.
This patch sets the visibility of a panel to INVISIBLE when it's
alpha is 0. This will ensure that a user cannot interact with a
view that cannot be seen. Not doing this resulted in a bug where
the user will tap on an area of the toolbar and activate an
unintended action.

It also removes the redundant mResetOverflowButtonDrawable. It's
actions are already performed in setPanelsStatesAtRestingPosition.

Bug: 27693433
Change-Id: Iafcde23622c7d1604924c520f6748fdbb8b135e7

Former-commit-id: 0cf648917fcbea6c91406566639294caacfa6fa1