Add error logging for softap commands
Log response code and message for softap commands to netd in failure
cases.

This CL adds a general function that logs an error if the executed
command does not return the expected response code or response message.
This function is general and can be used for commands other than softap.

BUG: 27834948
Change-Id: I88a5f78ccd5fc8d15af9726f2f59df0c27d49a4f

Former-commit-id: e2ec94fbab91e5da515617adb2d0532f7b52727f