Run cleanup before end actions of ViewPropertyAnimator.
Bug 27788719

The end action of a ViewPropertyAnimator may do anything,
including starting a new animator. If the next animator has
started prior to cleaning up the previous one, it will capture
the dirty state as the final state. This CL runs the cleanup
Runnables before running the end actions so that any captured
state is the clean state.

Change-Id: Ib005b817d420e79b636e61987669a852e15df9ce

Former-commit-id: e9a4f878528976efb4ec6f3d7797bef44732b00e