Fix IME adjustment for docked.
If we look in PhoneWindowManager offsetInputMethodWindow
lw we see mContentBottom will be updated to account
for the input methods appearance. Now if we look
at the non SOFT_INPUT_ADJUST_RESIZE case
in layoutWindowLw "normal window" (L4625)
we can see pf will be set to the adjusted mContentBottom
while cf will retain the non adjusted mDockBottom. So
adjustment by cf isn't sufficient we need to use pf. I've
preserved the existing behavior for freeform, until there
is more time to understand the desired behavior here, but
it seemed to make more sense for docked to reduce the available
area from the bottom rather than move the top offscreen.

Bug: 26387930
Change-Id: Id6cf70ea2b584939e24a12f9a32986c096f3d3eb

Former-commit-id: fc03b2bdb7596106b46075e5524514820faccbdd