Support multiple Enrollment APKs
Previously, the KeyphraseEnrollmentInfo would iterate through the list
of Enrollment APKs and just use the first one for a list of supported
keywords (of which there only could be one). Now, all APKs will be
enumerated and the enrollment APK that provides the given
keyphrase/locale will be used for the enrollment intent.

Change-Id: I78b0ef758c6c024b5787603bba5907a646f3c3f1

Former-commit-id: 1cba0696bd4c0e9d9d35f919a7a4c029ba9e4956