Ensure we leave space between layers in docked stack.
We need to leave space WINDOW_LAYER_MULTIPLIER between Z-order
adjacent normal windows, for things such as Dim Layers, etc. The
existing logic was breaking DIM_UNDER DimLayers in DockedStack. Consider
two normal windows, the second on top and with a DIM_UNDER. If we are
docked we will have restacked all the docked windows at HighestLayer+1
HighestLayer+2, etc...and so our second normal window is only one higher
than our first one. So when we put a dim under the second window
we end up with the same layer and the output is undefined.

Bug: 27721903
Change-Id: Ic27e0a10da2d1ab4a6058e0755ad4409b29527eb

Former-commit-id: 6412d755f4eb0b562c6d25ffb1ac5357105d6b1c