Properly handle DimLayers which become fullscreen.
If a DimLayer becomes fullscreen while it is animating out,
then it will not be the shared fullscreen dim layer. However,
state.continueDimming will be false (as we are ending the Dim),
so around DimLayerController.cpp L191 we will identify it as
fullscreen, skip animating it for now, and not set fullScreenAndDimming
to the index (because continueDimming is false). fullScreenAndDimming
will be -1 so we will now animate the dimmer at index "fullscreen"
this will simply be the last fullscreen state we found, which would be
fine if the state we had skipped was the shared fullscreen dim layer,
but it wasn't. Ensure that we identify the layer as fullscreen and the
shared fullscreen dim layer.

Bug: 27775427
Change-Id: I49d4f59bb09af6a7e5bd666e499be1c7e11ee766

Former-commit-id: 063988c40ca328b32198cf35fd99a1fd314ba3da