Don't broadcast <Standby> message during shutdown when setting is disabled
When the system is shutting down or the screen is going to be off, <Standby>
message will be broadcasted if and only if the setting variable,
Settings.Global#HDMI_CONTROL_AUTO_DEVICE_OFF_ENABLED, is enabled. This change
will make the shutdown case to be checked with the setting properly, too.

Bug: 27781559
Change-Id: Ia6605d6b5627ab1ff67285d72846b618a09548bc

Former-commit-id: 9ccff51b54ce282d19cb13dd326dc03950be07fe