Don't performPendingDeferredStart on fragments that aren't added
Framework edition

When setting a user visibility hint on a fragment it was possible to
have a FragmentManager already set due to a transaction being composed
but before the commit, so only checking that we have one isn't
sufficient for determining whether we should perform a pending
deferred start. This could happen easily with ViewPager adapters. Use
isAdded() instead.

Also fix a bug where we could set our internal tracking of fragments
that need a deferred start incorrectly.

Bug 27814550

Change-Id: I3be65728650c48d27a3a9dd4ca15f52006d14ab9

Former-commit-id: b963c99c3a1d896c45587c997e5293738c0a8a80