Fix flicker when unlocking
When the device was locked the fullscreen stack was laid out
fulllscreen, even though the configuration was half-screen, which
lead to a race between the app drawing and the unlock animation
so sometimes you could see the fullscreen frame when the unlock
animation started.

Instead, only layout in fullscreen if the docked stack is fully
"hidden".

Bug: 27154882
Change-Id: I4ba0c396eb0312c2bf2d911903b68c88d28aae8c

Former-commit-id: f4fa8cb79bc9102d4bb6d18698fd01dc038d9b81