Fix a bug that IMS#mShowInputFlags is never updated.
As a preparation to fix bug 26985193, this CL fixes a bug that
IMS#ShowInputFlags is never updated.  As a result,
IMS#onConfigurationChanged() has always called
IMS#onShowInputRequested() with specifying 0 to flags parameter, which
is wrong.

With this CL, we can assume that IMS#mShowInputFlags keeps tracking the
last value when it should do.  Also, we can remove IMS#mShowInputForced
since it is now determined by IMS#mShowInputFlags.

Bug: 26985193
Change-Id: Ieff634b69ce941d78ffe208547f7edd19ea28643

Former-commit-id: ef5b4657a8b03b33d6c4087e0ced33606fcf05f3