Split WebViewUpdateService logic into fallback logic and update-logic.
Add an inner class for the updating-logic to differentiate that logic
from the rest of the service (logic such as keeping track of
relro-creation and the current WebView implementation).
The fallback logic doesn't hold any state but instead just checks the
state of the system to determine whether to enable/disable the fallback
package and therefore that logic doesn't really need its own class.

Since the minimumVersionCode check stores state we here move that into
the inner class containing the updating-logic.

Bug: 27635535

Change-Id: I0afdb74c12139439ca66bfbd1c173ad98662ed31

Former-commit-id: 3a6e6b24fbdb7b181c34583af97306eec7d81ac8