Fix dismiss all clipping for tablet / landscape
Shouldn't animate the whole view, use the translatable views so that
they get clipped on tablet / landscape.

Additionally this changes the behavior in the case where the gear
might be showing, previously that would hide prior to the dismiss
all animation, UX prefers this to just have the notification cover
the gear and slide away or the gear fade away depending on the side
it's showing.

Bug: 28006639
Change-Id: I43027cb29410496de5b6e09c4a0c57a69484295c
Former-commit-id: 9c2c49634f0d457c1ab79772d3e817df629933df