Adds a BigPicAmbient flag to WearableExtender
This adds the BigPicAmbient flag to WearableExtender, demonstrating that
a BigPictureStyle on a Notification (pimary page or content page)
should be converted to low-bit black and white on entering ambient,
and should stick around during ambient instead of being closed. This
is mostly for QR and barcodes for situations akin to flights.

Bug:28042177
Change-Id: I84065ab3966301c063d393b8fa6451ec527944b9

Former-commit-id: 4bcb06b961a3984eed0029162291f1b3760815ec